server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.11.7
python-dotenv==1.1.1
tenacity==8.2.3
email-validator==2.1.0
